Implement sparse index support (#327)
## Problem
`2025-01` includes support for working with sparse-only indexes. The
types and client code need to be adjusted to support the flexibility of
working with `values` and/or `sparseValues` in records, and for creating
sparse and dense indexes.

## Solution
- Update and refactor `createIndex` and its validator to support working
with the variations of different indexes:
  - Serverless - sparse or dense
  - Pod - dense only
- Update `Record` to allow optionality on `values`. Currently, we need
to support sending `values`, `values & sparseValues`, or `sparseValues`.
- Clean up some of testing harness and validation logic. There were also
some test cases around index creation and upserting that weren't being
covered.

```typescript
import { Pinecone } from '@pinecone-database/pinecone';
const pc = new Pinecone();

// create sparse index
await pc.createIndex({
  name: 'sparse-index',
  metric: 'dotproduct',
  spec: {
    serverless: {
      cloud: 'aws',
      region: 'us-east-1',
    },
  },
  vectorType: 'sparse',
});

// upsert sparse vectors
await pc.index('sparse-index').upsert([{
    id: 'record-1',
    sparseValues: {
        indices: [0, 2, 3],
        values[0.3, 0.8, 0.3]
    }
}])
```

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Describe specific steps for validating this change.


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209100496456015
  - https://app.asana.com/0/0/1209238391918339
  - https://app.asana.com/0/0/1209238391918355
  - https://app.asana.com/0/0/1208897809680463
  - https://app.asana.com/0/0/1209238391918350
  - https://app.asana.com/0/0/1209238393587514
  - https://app.asana.com/0/0/1209238391918336
  - https://app.asana.com/0/0/1209238391918360

+++ src/assistant/data/deleteFile.ts
@@ -1,15 +1,6 @@
 import { DeleteFileRequest } from '../../pinecone-generated-ts-fetch/assistant_data';
 import { AsstDataOperationsProvider } from './asstDataOperationsProvider';
-
-/**
- * The `DeleteFile` interface describes the file ID for deleting a file uploaded to an Assistant.
- */
-export interface DeleteFile {
-  /**
-   * The ID of the file to delete.
-   */
-  fileId: string;
-}
+import { PineconeArgumentError } from '../../errors';
 
 /**
  * Deletes a file uploaded to an Assistant by ID.
@@ -35,11 +26,16 @@ export const deleteFile = (
   assistantName: string,
   apiProvider: AsstDataOperationsProvider
 ) => {
-  return async (options: DeleteFile) => {
+  return async (fileId: string): Promise<void> => {
+    if (!fileId) {
+      throw new PineconeArgumentError(
+        'You must pass the fileId of a file to delete.'
+      );
+    }
     const api = await apiProvider.provideData();
     const request = {
       assistantName: assistantName,
-      assistantFileId: options.fileId,
+      assistantFileId: fileId,
     } as DeleteFileRequest;
     return api.deleteFile(request);
   };
